Fix the locmax issues
