+++ example/lib/dialogs/file_metadata.dart
@@ -1,0 +1,80 @@
+/// A dialog to display file metadata information.
+///
+/// Copyright (C) 2024, Software Innovation Institute, ANU.
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License").
+///
+/// License: https://opensource.org/license/gpl-3-0.
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://opensource.org/license/gpl-3-0>.
+///
+/// Authors: Dawei Chen
+
+library;
+
+import 'package:flutter/material.dart';
+
+/// Shows a dialog displaying file metadata.
+
+void showFileMetadataDialog({
+  required BuildContext context,
+  required String fileName,
+  required String contentLength,
+  required String lastModified,
+  required String contentType,
+  required String allowdAccess,
+}) {
+  showDialog(
+    context: context,
+    builder: (_) => AlertDialog(
+      title: const Text('File Information'),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          _infoRow('File name', fileName),
+          _infoRow('Last modified', lastModified),
+          _infoRow('Content length', contentLength),
+          _infoRow('Content type', contentType),
+          _infoRow('Allowed operations', allowdAccess),
+        ],
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.pop(context),
+          child: const Text('Close'),
+        ),
+      ],
+    ),
+  );
+}
+
+Widget _infoRow(String label, String value) {
+  return Padding(
+    padding: const EdgeInsets.symmetric(vertical: 4),
+    child: Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(
+          width: 120,
+          child: Text(
+            '$label:',
+            style: const TextStyle(fontWeight: FontWeight.bold),
+          ),
+        ),
+        Expanded(child: Text(value)),
+      ],
+    ),
+  );
+}
